python-app/pyproject.toml
[project]
name = "python-app"
version = "0.1.0"
description = "app for connecting to llm"
authors = [
    {name = "dhyanesht",email = "t.dhyanesh@gmail.com"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "llama-index-llms-ollama (>=0.9.1,<0.10.0)",
    "fastapi (>=0.127.0,<0.128.0)",
    "uvicorn[standard] (>=0.40.0,<0.41.0)",
    "groq (>=1.0.0,<2.0.0)",
    "llama-index-llms-groq (>=0.4.1,<0.5.0)",
    "confluent-kafka (>=2.12.2,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

